wifi: Don't call wpa_supplicant from binder thread context
Commands to wpa_supplicant are serialized through WifiStateMachine.
This rule is broken when dumpstate collects data from the wifi
service. In this case a command will be sent wpa_supplicant from the
binder thread.

If wifi is turned off at the same time as dumpstate is running the
wpa control socket structure can be cleared leading to a crash
in system server.

Since the info from the dump is not of high value we can remove it.

Bug: 33066580

Change-Id: I4f829d017176261272b479cef9dcb0e10baca345