Wifi: Update WifiController state if Wifi turn On fails
If Wifi turn On fails due to failure in reloading firmware then
WifiStateMachine stays in InitialState, but WifiController is left
stuck in StaEnabledState, which in turn fails to turn on WLAN again.

Set Wifi state as failed to generate intent WIFI_STATE_CHANGED_ACTION so
that WifiController transitions to ApStaDisabledState.

Change-Id: I007d85179a0566e2f1b452ae8b564c4e8258987d