Refactor for accessor changes to WifiEnterpriseConfig
WifiEnterpriseConfig has changed to provide more sane accessors.
Now, callers are not given blanket access to the internal fields
and need to pass an accessor object to either save or restore
wpa_supplicant configuration.

This CL also adds unit tests to validate the new behavior of EAP
and Phase 2 parameters to ensure that recent changes do not cause
any regressions.

Bug:26400915
Change-Id: I575dc1688ff4b6055708e135b2692c982bc89d7a