WifiLogger: fix startLogging()
startLogging() asks the native layer for meta-data about the ring
buffers, and then, if meta-data was _not_ available, starts logging the
ring buffers.

That's pretty much the opposite of what we want. Perhaps even exactly
the opposite of what we want.

Invert the condition, to actually start the ring buffers.

BUG=26981286
TEST=unit tests

Change-Id: I3acf227a9b8d4a8970d9bdc7a52fefa101a32852