Move calculateWifiScore out of WSM and create new WifiScoreReport class
Pulled the calculation for a Wifi Score out of WSM and created a class to
hold score related objects.  This new class holds the score report
string along with the badLinkspeedcount int.  The calculateWifiScore
function is now a static method in the new WifiScoreReport class.  Added
the ScoreReport to a couple of places where related counters are also
cleaned up.

This CL also cleans up some style issues and converts hardcoded values
to static variables for the class.  No functionality changes were
introduced into the score calculation.

BUG=27312068

Change-Id: I167ecbef6e365923a687928235beecf3ac1b66f1