WifiStateMachine: Return to InitialState if driver start fails
If WiFi is set to shut off on idle, WifiController calls
mWifiStateMachine.setDriverStart(false), and WifiStateMachine
eventually enters the DriverStopped state.  On return from
idle the reverse happens and WifiStateMachine attempts to
restore the driver.  If this fails, WifiStateMachine simply
returns to the DriverStopped state.

This CL changes this behavior.  If restarting the driver
fails we now restart supplicant, which causes the state
machine to traverse the InitialState.  This provides a fail
safe mechanism for restarting drivers that may not have
suspended correctly.

Bug: 26735919
Change-Id: Ide9bc5339a9be1ea4ac67fc99fe6065529ee67e2
Test: Reproduce b/26735919 on bullhead, which is currently suffering from b/27576943