WifiScanningServiceImpl: Sw PNO scans
Moving the implementation of SW PNO scans from SupplicantScanner to
WifiScanningServiceImpl.

WifiScanningServiceImpl checks is HW PNO is supported by the
scanner implementation:
a) if false it manually schedules a background scan and
converts all background scan results to PNO results.
b) if true, it lets the scanner implementation handle the PNO scans.

BUG: 27742823
Change-Id: Ia15900d402ba005d11251931b39bf30b8d21a02c
TEST: Compiles & unit-test passes