WifiScanningService: Remove |pnoSettings| in stopPno
Remove the |pnoSettings| param for stopPno requests. We added this param
to revert network priorities in supplicant after PNO. But, because we
have disabled all auto-connect in wpa_supplicant having the wrong
priorities is irrelevant now.

Note: Will add a follow up CL to remove it from the WifiScanner
interface.

BUG: 27532350
Change-Id: I8bc4102b7e45699152a2ee0f816d65738b8531fa
TEST: Compiles & unit test passes