WifiScanningService: Create internal clientInfo
Create a new class for keeping track of internal clients to the
WifiScanningService. This object is used to communicate between the
state machines using a valid clientInfo.
The external clients override the onReportEvent to send the events back
to the WifiScanner interface.
The internal clients override the onReportEvent to send the events to
the state machine which requested the event.

Internal Clients are going to be used by WifiChangeStateMachine and the
upcoming WifiPnoScanStateMachine to make background scan requests

BUG: 27727025
Change-Id: Ic84bc39f59890151c51ac14356e1a3183819e09d
TEST: Compiles & unit-test passes