WifiStateMachine: reduce number of LogRecords
Allowing 3000 LogRecords uses too much memory (we've
seen >1 MB in some cases). Lower this limit to 100
records, as it was before.

To facilitate debugging issues that happen overnight,
allow the user to opt-in to a larger log record buffer.
To do so, the user should "enable wi-fi verbose logging"
under developer options.

BUG=27504679
TEST=unit tests

Change-Id: Iee983e3c815bef1a9e05f7085147bc73c79753f6