WifiConfigManager: instantiate mLocalLog unconditionally
Previously, mLocalLog was instantiated only if the
field mShowNetworks was true. Since this field was
always set to true, so the local log was always
instantiated.

Now that mShowNetworks has been replaced by sVDBG, which
defaults to false, mLocalLog no longer defaults to a non-null value.

Since we want WifiConfigManager to always use a local log,
instantiate mLocalLog unconditionally instead.

BUG: 28276051
Change-Id: I1694357acdf044390cd10f7e1a4a35fb58973f0d
TEST: 'runtest frameworks-wifi' passes.