NetworkDetail: remove VDBG field
The VDBG is field us currently used as the DBG field is
in other files: as a static, compile-time flag that
determines whether certain messages are logged. The DBG
field, on the other hand, is currently defined but not used.

Replace the uses of VDBG with DBG, and remove the VDBG field,
so that NetworkDetail will only have a single compile-time
debug flag that is consistent with those used in other WiFi
framework classes.

BUG: 28321397
Change-Id: I0b32f47e50e8add7e889ac72e7c9c259802bf083
TEST: 'runtest frameworks-wifi' passes.