Clear config BSSID after successful roaming
Clear the config BSSID after roaming completed and before entering
ConnectedState. Wifi chipset may roam from this point on and
having the BSSID specified by QNS would cause the roam to fail
and the device to disconnect.

We used to rely on the fact that entering ObtainingIpState
clears the config BSSID. But this assumption was broken by
commit fd3547987a12a808a046, which changed from transition
to ObtainingIpState to ConnectedState.

Bug: 28520770
Test: Wifi unit tests, manual tests
Change-Id: I09a2d5d1132ff597c63a324a24d13719be384603