[NAN] Reset current command when a timeout is received
The current command must be reset (to null) once a timeout is
received. Results in a WTF message being printed out (when a new
command is received) - though would not impact functionality since
any subsequent response would be received in the wrong state or
after a new command was being executed.

Change-Id: I1a62b3263b62647d67a1593c026b6bf5973c8486