[NAN] Create/delete all NAN data-path interfaces on Wi-Fi up/down
Creates (deletes) all NAN data-path interfaces when Wi-Fi comes up (down).
All NAN data-path is determined by the capabilities of the device. The
interfaces are created but are not up. They are only brought up when an
actual data-path is created.

Rationale for not creating on demand: latency will be too long - especially
for responder (which has an OTA response expectation).

Bug: 26562610
Change-Id: I090af999e0f251a062f216b6d39d29cda7635882