WifiController: Add CMD_RESET_WIFI command
Allow WifiController to safely restart the WiFi stack.  This will
initially be used by the WifiLastResortWatchdog.  The CMD_RESET_WIFI
message is only handled in the StaEnabledState where it would have been
attempting to connect.  The process of bringing the wifi back up is
handled through the use of the new CMD_RESET_WIFI_CONTINUE command.
Tests are also added to verify that the new CMD_RESET_WIFI command
is ignored in other states (explicitly tests Emergency mode and
AP enabled).

BUG: 27856267
Change-Id: I778ccd6f7d555f6ee6abb195c1c16c106c2e66b7