WifiStateMachine: Return null link layer stats
Modify |CMD_GET_LINK_LAYER_STATS| handling to return a null object when
the stats cannot be retrieved from wifi HAL. We've added a
feature check in |WifiServiceImpl.reportActivityInfo| method to check if
the device supports link layer stats reporting or not. So, there is no
need to fake an empty object if the retrieval fails.

BUG:29117407
Change-Id: Iaf8e060912611773398210785ebfaae7371db77c