WifiConfigStoreNew: Create skeletal class
Create the skeletal class for the new WifiConfigStore. The new
persistent store is composed from 2 separate files:
1. Shared store file to store all shared configurations/global data.
2. User specific store file to store all private configurations/user
data.
|StoreFile| sub class is a wrapper over the atomic files used to
write/read the raw data to/from the files.

|wifiConfigStoreData| class is used to pass the data to be written/read from
the config store files. This object is used to serialize/deserialize the
data to be stored in the store files.

Also, create a skeletal test class to redirect the read/write data from
the ConfigStore to an internal buffer for testing.

BUG: 29179278
Change-Id: I38dad84c0b2444e2f4ffe30658c8f3eea0127143
TEST: Unit tests