WifiConfigManager: move dumping of LocalLog
The dump() method of WifiConfigManager dumps the
LocalLog that's used by WCM. That would make sense,
except that the LocalLog that WCM uses isn't actually
owned by WCM.

Instead, the LocalLog is allocated by WifiNative, and
shared by multiple objects. (The sharing is safe, because
LocalLog's methods are synchronized.)

Since the LocalLog in question is a shared facility,
let's move dumping of the LocalLog to WifiLogger.

BUG=29424414
TEST=unit tests
TEST=manual

Manual test
$ adb bugreport foo.zip
$ unzip foo.zip
$ grep 'WifiNative LocalLog' bugreport*.txt
-> expect a match

Change-Id: I291ac29385472fd0e56fcb522d818dcffcedfe36