Fix tethering failure when wifi is not started
getCurrentCountryCode() returns the country code which we already
sent to driver.

However, sometimes when wifi and location scan are not started,
we have a valid country code but we didn't set it. In this case
getCurrentCountryCode() returns null. This will confuse SoftApManager
and break tethering. This also causes WifiApDialog not to show 5GHz
option in the UI when user creates a new wifi hotspot config.

In this CL we fix the incorrectly reported country code by adding a
function getCountryCode() which returns a valid country code regardless
of was it sent to driver or not.
SoftApManager and WifiApDialog will use this function to make decisions.

In addition, we also rename getCurrentCountryCode() to getCountryCode
SentToDriver() for better readability.

BUG=30200338
TEST=compile
TEST=unit tests
TEST=manual tests

Change-Id: I5ba576509f9f401f4d57a4628a147ac8871552c0