QNS: candidate network BSSID
QNS recommends a network for WCM to connect or roam to by
returning that network's configuration with the BSSID stored
in the 'candidate' field.

In a corner case that if 1) two APs are within the same network
(SSID), 2) the two APs get exactly the same QNS score (one AP has
higher RSSI value but 2.4GHz, and the other is 5GHz but lower RSSI),
and 3) the 5GHz AP shows up in the scan results before the 2.4GHz AP,
QNS sets the wrong BSSID value. This change fixes that problem.

Bug: 30184472
Change-Id: I8c67ab194571ba17b853eae617fb659017f72969
Test: WiFi Framework Unit Tests & manual tests