WifiService: add WifiLog interface
As a first step towards updating the logging
functionality in WifiService, we need to abstract
the existing liblog-style logging that we use.

To that end, add:
1) WifiLog, an Interface for logging, and
2) LogcatLog, an implementation of WifiLog,
   which uses android.util.Log as the backend.

Clients of android.util.Log, within WifiService,
will be migrated to this new API over time.
The first client to be migrated will be
WifiDiagnostics (in my very next CL).

BUG=30564891
TEST=wifitests/runtests.sh

Change-Id: Idd907c4bd8a2931d85b036593b6952018b46c6fb