move WifiDiagnostics creation from facade to injector
We're about to change the WifiDiagnostics ctor, to
take an extra argument. As long as we're going to be
mucking with the WifiDiagnostics ctor, let's move the
WifiDiagnostics construction call over to WifiInjector.
(WifiInjector is where kuh@ and all the cool kids
hang out these days).

Additional benefits of this change:
- The change slightly simplifes the WifiStateMachine
  constructor.
- The change makes the behavior of WifiStateMachineTest
  more uniform. Previously, the test would sometimes
  use a mock BaseWifiDiagnostics, and sometimes use a
  (real) WifiDiagnostics. Which path we took would depend
  on board-specific resource settings.

BUG=30566652
TEST=wifitests/runtests.sh

Change-Id: Iee374b9ef49db9264c544e666de5d4621defdfc7