WifiLog: add CheckReturnValue annotation
We'd like ErrorProne to flag misuse of the WifiLog API.
Add the @CheckReturnValue annotation, so that ErrorProne
can help us out.

Note that the @CheckReturnValue annotation requires adding
jsr305 to the WifiService build. This is not, however, a
new dependency for the framework in general. In particular,
jsr305 is already used by frameworks/base/packages/EasterEgg.

While there: fix-up use of @ThreadSafe in LogcatLog. (Now
that we're pulling in jsr305, we can actually use the
annotation, instead of having the annotation commented out.)

BUG=30738484
TEST=wifitests/runtests.sh
TEST=mma -j32 RUN_ERROR_PRONE=true javac-check (might fail)
TEST=find . -name '*.java' | xargs touch
TEST=mm -j32 RUN_ERROR_PRONE=true javac-check (should pass)

Change-Id: I5d55b287b197710cde3fa0d31c8a5ccf09edd2fa