WifiConfigManagerNew: Implement channel set creation
Implement the channel list creation for the currently connected network
for partial scans. The channel list is created using the scan detail
cache for that network and any linked networks. Scan results which are
stale beyond the provided the age are ignored during channel list
creation.

While there,
1. Set the |ScanResult.seen| parameter when ScanDetail is created. This
was a pending todo from b/29877741 (See comment#3) which we wanted to do
in master branch.
2. Truncate the length of the SSID base used to create SSID's in
|WifiConfigurationTestUtil|. This was exceeding the max size of SSID
(32) when the number of networks created during tests exceed 99.

BUG: 30156343
TEST: Added Unit tests
Change-Id: I3e2a18ba72c9c244235971fc29e0711526510295