Always start/stop wpa_supplicant service on request
For a given product, P2P is either supported or not supported,
depending on the product configuration and hardware support.
Because of this, on a given product, we do not need to determine
which wpa_supplicant configuration to run, we can just use "whatever
the given product supports."

We've changed init files to define a single service "wpa_supplicant".
In places where there were both p2p_supplicant and wpa_supplicant
services defined, we have removed the wpa_supplicant service, and
replaced it with the definition of p2p_supplicant.

Above the supplicant we choose for a given platform, we still keep
the hooks that allow optional P2P support in the framework, since
the framework needs to support hardware with P2P support, and hardware
without.

Bug: 30816535
Test: unit tests pass on bullhead
Test: cts-tradefed run cts-dev \
          --module=CtsNetTestCases \
          --test=android.net.wifi.cts.ConcurrencyTest passes

Change-Id: I6ea65fde1707f65e70b41c94ca823bb2318d81f0