SupplicantWifiScannerImpl: Restart PNO on settings change
When a new network is added, ConnectivityManager will stop/restart PNO
scans. However because of the 5 second debounce logic in HwPnoDebouncer,
this toggle will not be sent to wpa_supplicant. So, force restart PNO in
SupplicantWifiScannerImpl whenever there is a change in the network list .

PS: wpa_supplicant already has some internal logic to handle network
additions when PNO scans are running. But, not sure if this is working
as intended.

BUG: 30833790
Change-Id: Ib328abd866298a8a672a2f19fc863ee515dcbe74
TEST: Added unit tests