WifiScanningService: Move scan results bcast
A WifiManager.SCAN_RESULTS_AVIALABLE_ACTION broadcast is sent when scan results
are returned for an app/user initiated scan.  This broadcast is now
located in the WifiScanningService instead of WifiStateMachine.  Further
cleanup will move retrieval of the scan results from WifiStateMachine to
WifiScanningService.  Tests were added to WifiScanningServiceTest to
verify the broadcast is sent in both scanning success and failure cases.

BUG: 31347837
TEST: frameworks/opt/net/wifi/tests/wifitests/runtests.sh on angler
Change-Id: I7c11d571552090057593539b0ed736bc2cedeb00