WifiStateMachine: Send DHCP_ACTION_COMPLETE if p2p is missing
With ag/1066617, null checks were added for mWifiP2pChannel
in WifiStateMachine.  An issue was created during DHCP setup where
in this null condition the CMD_PRE_DHCP_ACTION_COMPLETE message is
never delivered, since the original code depended on WifiP2pService
to deliver it.  In the null case, we take responsibility to deliver
this message directly from WifiStateMachine.

Bug: 32621474
Change-Id: I7814d089479cbbb9c2407e782578a09e4c823b08
Test: Manual: remove mWifiP2pChannel initialization in InitialState.enter