WifiConfigStore: Log the exception in buffered write
I suspect the store file write failures reported via wtf are due to
the user's CE store not being visible (device is still booting up). This
should be pretty harmless because we would eventually write out the
data when the user unlocks. So, add the exception reason to the wtf to
confirm the root cause before putting in a fix.

Bug: 33446122
Test: Compiles
Change-Id: Iff6bef5acc4e55067169c3e2b74500df28c4e694