WifiStateMachine: check p2p supported
WifiStateMachine communicates with the P2P service without checking if
P2P is supported by the device.  This CL adds a check for p2p support
and only attempts to create an async channel with the p2p service if
it is supported.  If it is not supported, the p2p service and p2p
async channel local variables will be null and no messages will be
sent between services.

Bug: 34193861
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: frameworks/base/wifi/tests/runtests.sh
Test: manually checked p2p peers
Change-Id: Ie4d22aff8de02086142371fbe7da705e669921ab