WifiConfigManager: Restructure config store load
Currently, we load from config store only when we start wpa_supplicant
in SupplicantStartingState. However, we handle the user unlock/switch
in the DefaultState. This could lead to cases where we handle the user
unlock (& thus overwrite the config store file) even before we have had
the opportunity to read the existing data out.

Changes in the CL:
1. Decouple the migration of data out of legacy store and the load of
data from new store files.
2. The new config store loading will now happen in BOOT_COMPLETED
message handling. If it's a bootup after fresh install or upgrade from
N, this will be a no op.
3. Whenever wpa_supplicant is started, attempt to migrate data out of
the old config store. If the data has already been migrated, this will
be a no op.
4. User unlock/switch before migration is performed is ignored.

PS: This will also help us handle the
|WifiManager#getConfiguredNetworks| API even if wifi is disabled.

Bug: 33966771
Bug: 33391502
Test: Added unit tests
Test: Will send for integration tests
Change-Id: I3e4f7e21dc88bad5ae8581a150b29850fbac0afa