Apply client chain when installing enterprise credentials
Extract the client keychain and install this as a single blob
after the client certificate within the blob.  Doing so is
backwards-compatible since wpa_supplicant will terminate at the
end of the client certificate at the end of the first
"PEM_read_bio_X509()" call when this blob is retrieved.  If
wpa_supplicant knows how, it can then retrieve the rest of the
keychain using successive PEM_read_bio_X509() calls on the same
input data.

Bug: 34688653
Test: Regression tests using WiFi testbed
Change-Id: I9251a4da59ca189a8682c632966fc305c1ec126e