hotspot2: PasspointProvider: update semantics for certificate and key alias
Instead of using the full name for the installed certificate and key as the
alias, use the suffix of the name (without the certificate or key type).
This will make the certificate/key alias meaning consistent with the usage in
WifiEnterpriseConfig.

This will make it cleaner/easier for converting the legacy Passpoint configuration
to the new one.

Bug: 34206769
Test: frameworks/opt/net/tests/wifitests/runtests.sh
Test: manual testing using a Boingo profile
Change-Id: Iab005d98f83c9e217f87ab780787bdcc618b6509