hotspot2: update semantics for PasspointManager#matchProvider
Instead of returning a list of matched providers, it will now
return the best provider for the given AP (based on ScanResult).
It will return null if no match is found.

So PasspointManager is responsible for ranking providers for
a given AP (based on ScanResult), while PasspointNetworkEvalutor
is responsible for ranking Passpoint networks with matched provider.

Bug: 35888100
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: Manual test with Global Reach Passpoint AP
Change-Id: I9e2a76aaf319f6180d7767fe8f713547221feb6c