Only auto-disqualify 2.4GHz NWs when 5GHz NWs are available
The WifiSelector logic automatically disqualifies 2.4GHz
networks so that we always go looking for a candidate 5GHz
network. This is the right thing to do so long as there are
5GHz candidates. For 5GHz candidates to be available:
1) 5Ghz must be supported by the device
2) 5GHz networks must be found by a scan
3) 5GHz must be enabled by the user (not always configurable)

Thus, for cases where 5GHz networks aren't available for
whatever reason, we should only disqualify 2.4GHz networks
using the established RSSI-based logic that is applied in
other cases.

Bug: 36226123
Test: test2GhzQualifiedNo5GhzAvailable,
      test2GhzHighQuality5GhzAvailable,
      test2GhzNotQualifiedOther2GhzAvailable,
      test5GhzNotQualifiedLowRssi,
      test5GhzQualified
Merged-In: I3632e056c9890c7ac6607f72b21e42d34ec0386d
Change-Id: I3632e056c9890c7ac6607f72b21e42d34ec0386d