hotspot2: add support for the build config flag "config_wifi_hotspot2_enabled"
This flag specifies if the Passpoint is enabled or not.  Currently,
this flag is set to false for the default overlay.

So do not register PasspointNetworkEvaluator if Passpoint support
is not enabled to avoid any Passpoint related traffics (e.g. ANQP).
Also throw UnsupportedOperationException for any Passpoint related API
calls.

Bug: 32671424
Test: manual test
Change-Id: Icefa9b990d6d989750e5a7f00b4f3f5d83a82960