hotspot2: verify CA certificate before installing Passpoint provider
For Hotspot 2.0 Release 1, the provisioning method is not standardized,
so to improve security, the CA certificate must be trusted by one
of the pre-loaded public CAs in the system key store.

So verify the CA certificate if one is provided before installing
the Passpoint provider.

A utility class CertificateVerifier is created in order to easily mock
the certificate verification in unit tests.

Bug: 34460350
Test: manual test
Change-Id: I627ef9da2876ffbaf29aadb2d5b281c75ec1d145