WSM: Reconnect on network credential change
Currently, we ignore connect request if we're already connected to the
same network. This is not correct for network modifications. Hence,
trigger a reconnect in this case.

Changes in the CL:
1. Add a new |hasCredentialChanged| param in NetworkUpdateResult.
2. Use that in WSM to determine if we need to trigger a reconnect.

While there,
Removed a couple of unused setters in NetworkUpdateResult.
Fixed WifiConfigManager unit test failure due to mockito change.

Bug: 36505419
Bug: 36020928

Test: Unit tests and manual tests via settings UI.
Test: Will send for regression tests.
Change-Id: I1adaac58108c2f65fb36173a468dff000f5372c8
Merged-In: I1adaac58108c2f65fb36173a468dff000f5372c8