WifiConfigManager: Persist immediately on enable/disable network
The change in network state due to public API calls should be
immediately persisted (not wait for the buffered write from
updateNetworkSelectionStatus to kick in after 10 seconds).

This was detected in the test: |test_reboot_configstore_reconnect| which
connects to a network and immediately reboots the device. This prevents
the last buffered write (for enabling network) to be flushed out and
hence the network is reloaded as disabled.

Bug: 35811529
Test: Unit tests
Change-Id: I6f769429df67d83ad014ef552901a7b1419c7721