Remove networks from supplicant before starting PNO scan
We only start PNO scan when network is disconnected and screen
is off.
In the case we get disconnected from an AP, supplicant will keep
triggering scan and reconnect to that network.
This behavior is useful for immediate reconnection. However, when
PNO scan is on, the supplicant's single scan and PNO scan would
usually fail each other because they share the same radio resource.
This is causing power regression.
With this patch, we remove networks from supplicant so supplicant
will not keep scanning for reconnection when PNO scan is on.

Bug: 37744179
Test: unit tests, manual test, integration test

Change-Id: I03d4dea3b7f1acb51d716126f1057c13cd3ee9be