WifiServiceImpl: add binder death callback impl
Binder objects for requesting applications are tracked in
LocalOnlyHotspotRequestInfo objects.  When a binder death is detected,
this object calls back in to WifiServiceImpl so the request can be
removed and the hotspot can shutdown if there are not any outstanding
requests.

Added a test to verify behavior when there are not any pending requests,
further testing will be added when we start tracking the startLOHS
requests.

Bug: 31466854
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: Ida9cf996683acca03e5478845721501d12ee1a7f