hotspot2: PasspointNetworkEvaluator: update network candidate for current network
When evaluating ScanResults while connected, PasspointNetworkEvaluator
might decide to stay with the current active network based
on the ScanResults and installed providers.  In this scenario,
we need to update the selected BSS info for the current active
network, since it is possible to have multiple BSSes associated
with a given network.  It is up to the caller to decide if it wants
to initiate a new connection when new BSS is selected.

Bug: 38031353
Test: manual test
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: I9c43edae23264d779dfa34809590aeae12685faf