Do not remove and add the same network
This patch allows SupplicantStatIfaceHal to skip the "remove
and add" process for supplicant network if there is already
the same network saved in supplicant.

This will be helpful when phone reconnects to a EAP-SIM
network. In that case, EAP state parameters will not be cleared by
removing network, so supplicant can take advantage of those
parameters for fast-reauth.

Bug: 37529849
Test: compile. unit test, integration test, manual test

Change-Id: I4326e9f60f4f8a64674a52d1dbbcd34c5e709064