[AWARE] Wait for disable
The NAN disable command now waits for a response from the HAL. This
introduces a slight ordering in disable/enable sequences - which can
prevent some driver/firmware race conditions. Though it is not perfect
in that we are not waiting for the down indication (the state machine
isn't set up for that).

Intended to work-around a system crash regression observed by running
Cts SingleDeviceTest#testAttachDiscoveryAddressChanges which executes
rapid enable/disable cycles. Does not solve issue - but a good safety.

Bug: 38195097
Test: unit tests and integration (sl4a) pass
Change-Id: Idc4622ca69e7e1c1591a8bb1f6872f2d071c6e84