[AWARE] Inform app of attach failure when aware isn't available
When Aware is disabled there's a broadcast informing apps. This
also implies to app that all operations they have pending will
fail - whether or not they get callbacks for them (the whole service
is effectively shut down).

This change adds a failure callback in such a scenario for attach
create. Minor change which makes the interface a bit richer (app
doesn't have to coordinate two different information streams), and
is useful for test scripts validating such a flow.

Bug: 31940045
Test: unit tests + integration (sl4a) tests passing
Change-Id: I76985cc0ec24150bf3f6128e7f37e40c243ea717