WifiServiceImpl: implement stopLOHS
When a requestor is done with LOHS, closing the callback will trigger
stopLocalOnlyHotspot.  This CL implements the logic to unregister the
request and stop softap mode - if that caller was the last registered
requestor.

In this CL, LocalOnlyHotspotRequestInfo was also updated to use a
different key to track the requests.  This was done to support apps
calling with shared uids.

Bug: 31466854
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: I585ec1cbd4284e40e068bb47d12f80fc445d9270