WifiServiceImpl: listen and act on softap changes
The registered callers of LocalOnlyHotspot need to be updated when the
hotspot fails or stops.  This is currently implemented by having
WifiServiceImpl listen for WIFI_AP_STATE_CHANGE broadcasts.  When
updates about softap failures or stopping (disabling or disabled) are
received, appropriate messages are constructed and sent to registered
callers.  This CL additionally adds tests confirming this behavior.  An
additional helper methods were added to aid testing for sending the
broadcasts and sending messages to the registered LOHS requestors.

Bug: 31466854
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: I5ade861423c7ea25690a084e9e4c7c9861f1dbcd