WifiStateMachine: Reset network ID for WPS
|WifiNative.migrateNetworksFromSupplicant| retrieves the network ID of
the configuration from wpa_supplicant. This is useful for migration of
config store from N to O, but we don't need this network ID information
for WPS operations since the network ID's allocated by wpa_supplicant
will clash with network ID's allocated by framework. So, reset the
network ID before we add it to WifiConfigManager.

Bug: 62156507
Test: Unit tests
Change-Id: I1c24c60dc2de2d039b3c372d52d10d55cc4d0241