[AWARE] Delete all possible NDIs - not just those framework thinks exists
Currently The Aware framework keeps track of interfaces it created
and deletes those. If for some reason it loses track of a created
interface it will never delete it - stuck.

New: Just as for creating interfaces - where uses capability to create
all possible interfaces, use the same mechanism to delete all possible
interfaces - whether or not they exist. If they don't exist then will
get a failure (which we ignore).

Bug: 62276043
Test: no instance of left-around interfaces observed
Change-Id: Icb5f1f792448d2b16885645ca2ade4cd462b162e