Permanent disable network with wrong password
When connecting to a network for the first time, an indication
of "wrong password" from supplicant is a strong indication that
it is a wrong password.  We will permanently disable such network
in this case to avoid any retries and allow us to prompt the user
immediately. The network will be re-enabled when the user manually
trigger the connection again (most likely with a new password).

In the case when the network had been connected before, we are
assuming that we are attempting to connect to a wrong network (same
SSID different password).  Thus, the failure is assumed to be temporary.
The other scenario is when the user actually changed the password
on the AP, which is an unlikely scenario, so it is being ignored for now.

Bug: 62444625
Test: manually connect to a WPA network with wrong password, verify
      error message is display after first try and no more retries
Test: integration test

Change-Id: I2ab543e5b63546744291e87ec9d083e115aa9931