[AWARE] Add debug hooks to obtain count of event callbacks
Count Aware event callbacks and provide a shell command to
retrieve and reset data.

While callbacks may be percolated up through the framework and
detected - that is not always the case. Some of the callbacks are
masked by the framework. It is still interesting to know when
callbacks occur - to detect invalid wakeups of the host.

Bug: 31940045
Test: manual shell cmd execution validates data
Change-Id: I35d96419342c330f2d64a422b9b39369311aa09e