WifiStateMachineTest: Use mock IpManager
All we need from IpManager object in WifiStateMachineTest is the
IpManagerCallback object which can be used to inject events. So, instead
of creating a TestIpManager, return a mock IpManager and just capture the
incoming callback in the test for injection.

Usage of the real IpManager makes these unit tests flaky.

Bug: 62842968
Test: Unit tests run reliably now.
Change-Id: I3345bf6e101c6502c92fd2785082cfbe1084cefc