WifiConfigManager: Remove transient networks from PNO list
The network configuration (WifiConfiguration) objects created for
ephemeral/passpoint networks are valid only for the duration of a
particular connection attempt. These are generally deleted every time we
disconnect from such a network. However, there are cases where such
transient network configs linger on in the internal database after a
connection failure, for example. So, ignore such networks in the PNO list
because we don't want the device to wakeup for attempting connection
to such networks.

PS: Such transient network configs will eventually be removed from the
internal database on the next disconnection (Using a lazy removal
approach).

Bug: 62841520
Test: Unit tests
Test: Will send for regression tests.
Change-Id: I8265c32a4195e75903affc63dc70804237ce2dd8